+++ docs/examples/09-api-to-csv2.php
@@ -1,0 +1,10 @@
+<?php
+
+require_once __DIR__ . "/.init.php";
+
+$chainProcessor = getChainProcessor(__FILE__);
+
+$chainProcessor->process(
+    new ArrayIterator([["id"=> 1],["id"=> 4],["id"=> 7],["id"=> 10]]),
+    []
+);